Cacheverwaltung in der Administration ergänzt

+++ core/classes/System/Cache.php
@@ -16,6 +16,7 @@ class Cache {
 	public static $cache_dir = '/core/cache/';
 	
 	public static function loadCache($key,$ttl="no_ttl"){
+		if(Config::get('enable_cache') != "1") return;
 		$file = md5($key);
 		$path = $_SERVER['DOCUMENT_ROOT'] . self::$cache_dir . $file;
 		if(is_file($path)){
@@ -33,6 +34,7 @@ class Cache {
 	}
 	
 	public static function saveCache($key, $data){
+		if(Config::get('enable_cache') != "1") return;
 		$file = md5($key);
 		$data = serialize($data);
 		file_put_contents($_SERVER['DOCUMENT_ROOT'] . self::$cache_dir . $file, $data);
@@ -41,7 +43,9 @@ class Cache {
 	public static function clearCache($key){
 		$file = md5($key);
 		$path = $_SERVER['DOCUMENT_ROOT'] . self::$cache_dir . $file;
-		unlink($path);
+		if(is_file($path)){
+			unlink($path);
+		}
 	}
 	
 }
